+++ BossMod/Modules/Endwalker/Ultimate/DSW1/HyperdimensionalSlash.cs
@@ -97,7 +97,7 @@ namespace BossMod.Endwalker.Ultimate.DSW1
             }
         }
 
-        public override void OnEventIcon(BossModule module, uint actorID, uint iconID)
+        public override void OnEventIcon(BossModule module, ulong actorID, uint iconID)
         {
             if ((IconID)iconID == IconID.HyperdimensionalSlash)
             {
